Crate fuser

FUSE userspace library implementation

This is an improved rewrite of the FUSE userspace library (lowlevel interface) to fully take advantage of Rust’s architecture. The only thing we rely on in the real libfuse are mount and unmount calls which are needed to establish a fd to talk to the kernel driver.

Modules

consts

Structs

BackgroundSession

The background session data structure

FileAttr

File attributes

KernelConfig

Configuration of the fuse kernel module connection

ReplyAttr

Attribute Reply

ReplyBmap

Bmap Reply

ReplyCreate

Create reply

ReplyData

Data reply

ReplyDirectory

Directory reply

ReplyDirectoryPlus

DirectoryPlus reply

ReplyEmpty

Empty reply

ReplyEntry

Entry reply

ReplyIoctl

Ioctl Reply

ReplyLock

Lock Reply

ReplyLseek

Lseek Reply

ReplyOpen

Open Reply

ReplyStatfs

Statfs Reply

ReplyWrite

Write Reply

ReplyXattr

Xattr reply

Request

Request data structure

Session

The session data structure

SessionUnmounter

A thread-safe object that can be used to unmount a Filesystem

Enums

FileType

File types

MountOption

Mount options accepted by the FUSE filesystem type See ‘man mount.fuse’ for details

TimeOrNow

Possible input arguments for atime & mtime, which can either be set to a specified time, or to the current time

Constants

FUSE_ROOT_ID

Traits

Filesystem

Filesystem trait.

Reply

Generic reply trait

Functions

mountDeprecated

Mount the given filesystem to the given mountpoint. This function will not return until the filesystem is unmounted.

mount2

Mount the given filesystem to the given mountpoint. This function will not return until the filesystem is unmounted.

spawn_mountDeprecated

Mount the given filesystem to the given mountpoint. This function spawns a background thread to handle filesystem operations while being mounted and therefore returns immediately. The returned handle should be stored to reference the mounted filesystem. If it’s dropped, the filesystem will be unmounted.

spawn_mount2

Mount the given filesystem to the given mountpoint. This function spawns a background thread to handle filesystem operations while being mounted and therefore returns immediately. The returned handle should be stored to reference the mounted filesystem. If it’s dropped, the filesystem will be unmounted.