Trait fuser::Filesystem

pub trait Filesystem {

    fn init(
        &mut self, 
        _req: &Request<'_>, 
        _config: &mut KernelConfig
    ) -> Result<(), c_int> { ... }

    fn destroy(&mut self) { ... }

    fn lookup(
        &mut self, 
        _req: &Request<'_>, 
        _parent: u64, 
        _name: &OsStr, 
        reply: ReplyEntry
    ) { ... }

    fn forget(&mut self, _req: &Request<'_>, _ino: u64, _nlookup: u64) { ... }

    fn getattr(&mut self, _req: &Request<'_>, _ino: u64, reply: ReplyAttr) { ... }

    fn setattr(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _mode: Option<u32>, 
        _uid: Option<u32>, 
        _gid: Option<u32>, 
        _size: Option<u64>, 
        _atime: Option<TimeOrNow>, 
        _mtime: Option<TimeOrNow>, 
        _ctime: Option<SystemTime>, 
        _fh: Option<u64>, 
        _crtime: Option<SystemTime>, 
        _chgtime: Option<SystemTime>, 
        _bkuptime: Option<SystemTime>, 
        _flags: Option<u32>, 
        reply: ReplyAttr
    ) { ... }

    fn readlink(&mut self, _req: &Request<'_>, _ino: u64, reply: ReplyData) { ... }

    fn mknod(
        &mut self, 
        _req: &Request<'_>, 
        _parent: u64, 
        _name: &OsStr, 
        _mode: u32, 
        _umask: u32, 
        _rdev: u32, 
        reply: ReplyEntry
    ) { ... }

    fn mkdir(
        &mut self, 
        _req: &Request<'_>, 
        _parent: u64, 
        _name: &OsStr, 
        _mode: u32, 
        _umask: u32, 
        reply: ReplyEntry
    ) { ... }

    fn unlink(
        &mut self, 
        _req: &Request<'_>, 
        _parent: u64, 
        _name: &OsStr, 
        reply: ReplyEmpty
    ) { ... }

    fn rmdir(
        &mut self, 
        _req: &Request<'_>, 
        _parent: u64, 
        _name: &OsStr, 
        reply: ReplyEmpty
    ) { ... }

    fn symlink(
        &mut self, 
        _req: &Request<'_>, 
        _parent: u64, 
        _name: &OsStr, 
        _link: &Path, 
        reply: ReplyEntry
    ) { ... }

    fn rename(
        &mut self, 
        _req: &Request<'_>, 
        _parent: u64, 
        _name: &OsStr, 
        _newparent: u64, 
        _newname: &OsStr, 
        _flags: u32, 
        reply: ReplyEmpty
    ) { ... }

    fn link(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _newparent: u64, 
        _newname: &OsStr, 
        reply: ReplyEntry
    ) { ... }

    fn open(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _flags: i32, 
        reply: ReplyOpen
    ) { ... }

    fn read(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _offset: i64, 
        _size: u32, 
        _flags: i32, 
        _lock_owner: Option<u64>, 
        reply: ReplyData
    ) { ... }

    fn write(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _offset: i64, 
        _data: &[u8], 
        _write_flags: u32, 
        _flags: i32, 
        _lock_owner: Option<u64>, 
        reply: ReplyWrite
    ) { ... }

    fn flush(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _lock_owner: u64, 
        reply: ReplyEmpty
    ) { ... }

    fn release(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _flags: i32, 
        _lock_owner: Option<u64>, 
        _flush: bool, 
        reply: ReplyEmpty
    ) { ... }

    fn fsync(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _datasync: bool, 
        reply: ReplyEmpty
    ) { ... }

    fn opendir(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _flags: i32, 
        reply: ReplyOpen
    ) { ... }

    fn readdir(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _offset: i64, 
        reply: ReplyDirectory
    ) { ... }

    fn readdirplus(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _offset: i64, 
        reply: ReplyDirectoryPlus
    ) { ... }

    fn releasedir(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _flags: i32, 
        reply: ReplyEmpty
    ) { ... }

    fn fsyncdir(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _datasync: bool, 
        reply: ReplyEmpty
    ) { ... }

    fn statfs(&mut self, _req: &Request<'_>, _ino: u64, reply: ReplyStatfs) { ... }

    fn setxattr(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _name: &OsStr, 
        _value: &[u8], 
        _flags: i32, 
        _position: u32, 
        reply: ReplyEmpty
    ) { ... }

    fn getxattr(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _name: &OsStr, 
        _size: u32, 
        reply: ReplyXattr
    ) { ... }

    fn listxattr(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _size: u32, 
        reply: ReplyXattr
    ) { ... }

    fn removexattr(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _name: &OsStr, 
        reply: ReplyEmpty
    ) { ... }

    fn access(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _mask: i32, 
        reply: ReplyEmpty
    ) { ... }

    fn create(
        &mut self, 
        _req: &Request<'_>, 
        _parent: u64, 
        _name: &OsStr, 
        _mode: u32, 
        _umask: u32, 
        _flags: i32, 
        reply: ReplyCreate
    ) { ... }

    fn getlk(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _lock_owner: u64, 
        _start: u64, 
        _end: u64, 
        _typ: i32, 
        _pid: u32, 
        reply: ReplyLock
    ) { ... }

    fn setlk(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _lock_owner: u64, 
        _start: u64, 
        _end: u64, 
        _typ: i32, 
        _pid: u32, 
        _sleep: bool, 
        reply: ReplyEmpty
    ) { ... }

    fn bmap(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _blocksize: u32, 
        _idx: u64, 
        reply: ReplyBmap
    ) { ... }

    fn ioctl(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _flags: u32, 
        _cmd: u32, 
        _in_data: &[u8], 
        _out_size: u32, 
        reply: ReplyIoctl
    ) { ... }

    fn fallocate(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _offset: i64, 
        _length: i64, 
        _mode: i32, 
        reply: ReplyEmpty
    ) { ... }

    fn lseek(
        &mut self, 
        _req: &Request<'_>, 
        _ino: u64, 
        _fh: u64, 
        _offset: i64, 
        _whence: i32, 
        reply: ReplyLseek
    ) { ... }

    fn copy_file_range(
        &mut self, 
        _req: &Request<'_>, 
        _ino_in: u64, 
        _fh_in: u64, 
        _offset_in: i64, 
        _ino_out: u64, 
        _fh_out: u64, 
        _offset_out: i64, 
        _len: u64, 
        _flags: u32, 
        reply: ReplyWrite
    ) { ... }

}

Filesystem trait.

This trait must be implemented to provide a userspace filesystem via FUSE. These methods correspond to fuse_lowlevel_ops in libfuse. Reasonable default implementations are provided here to get a mountable filesystem that does nothing.

Provided methods

fn init(
&mut self,
_req: &Request<'_>,
_config: &mut KernelConfig
) -> Result<(), c_int>

Initialize filesystem. Called before any other filesystem method. The kernel module connection can be configured using the KernelConfig object

fn destroy(&mut self)

Clean up filesystem. Called on filesystem exit.

fn lookup(
&mut self,
_req: &Request<'_>,
_parent: u64,
_name: &OsStr,
reply: ReplyEntry
)

Look up a directory entry by name and get its attributes.

fn forget(&mut self, _req: &Request<'_>, _ino: u64, _nlookup: u64)

Forget about an inode. The nlookup parameter indicates the number of lookups previously performed on this inode. If the filesystem implements inode lifetimes, it is recommended that inodes acquire a single reference on each lookup, and lose nlookup references on each forget. The filesystem may ignore forget calls, if the inodes don’t need to have a limited lifetime. On unmount it is not guaranteed, that all referenced inodes will receive a forget message.

fn getattr(&mut self, _req: &Request<'_>, _ino: u64, reply: ReplyAttr)

Get file attributes.

fn setattr(
&mut self,
_req: &Request<'_>,
_ino: u64,
_mode: Option<u32>,
_uid: Option<u32>,
_gid: Option<u32>,
_size: Option<u64>,
_atime: Option<TimeOrNow>,
_mtime: Option<TimeOrNow>,
_ctime: Option<SystemTime>,
_fh: Option<u64>,
_crtime: Option<SystemTime>,
_chgtime: Option<SystemTime>,
_bkuptime: Option<SystemTime>,
_flags: Option<u32>,
reply: ReplyAttr
)

Set file attributes.

fn readlink(&mut self, _req: &Request<'_>, _ino: u64, reply: ReplyData)

Read symbolic link.

fn mknod(
&mut self,
_req: &Request<'_>,
_parent: u64,
_name: &OsStr,
_mode: u32,
_umask: u32,
_rdev: u32,
reply: ReplyEntry
)

Create file node. Create a regular file, character device, block device, fifo or socket node.

fn mkdir(
&mut self,
_req: &Request<'_>,
_parent: u64,
_name: &OsStr,
_mode: u32,
_umask: u32,
reply: ReplyEntry
)

Create a directory.

fn unlink(
&mut self,
_req: &Request<'_>,
_parent: u64,
_name: &OsStr,
reply: ReplyEmpty
)

Remove a file.

fn rmdir(
&mut self,
_req: &Request<'_>,
_parent: u64,
_name: &OsStr,
reply: ReplyEmpty
)

Remove a directory.

fn symlink(
&mut self,
_req: &Request<'_>,
_parent: u64,
_name: &OsStr,
_link: &Path,
reply: ReplyEntry
)

Create a symbolic link.

fn rename(
&mut self,
_req: &Request<'_>,
_parent: u64,
_name: &OsStr,
_newparent: u64,
_newname: &OsStr,
_flags: u32,
reply: ReplyEmpty
)

Rename a file.

fn link(
&mut self,
_req: &Request<'_>,
_ino: u64,
_newparent: u64,
_newname: &OsStr,
reply: ReplyEntry
)

Create a hard link.

fn open(&mut self, _req: &Request<'_>, _ino: u64, _flags: i32, reply: ReplyOpen)

Open a file. Open flags (with the exception of O_CREAT, O_EXCL, O_NOCTTY and O_TRUNC) are available in flags. Filesystem may store an arbitrary file handle (pointer, index, etc) in fh, and use this in other all other file operations (read, write, flush, release, fsync). Filesystem may also implement stateless file I/O and not store anything in fh. There are also some flags (direct_io, keep_cache) which the filesystem may set, to change the way the file is opened. See fuse_file_info structure in <fuse_common.h> for more details.

fn read(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_offset: i64,
_size: u32,
_flags: i32,
_lock_owner: Option<u64>,
reply: ReplyData
)

Read data. Read should send exactly the number of bytes requested except on EOF or error, otherwise the rest of the data will be substituted with zeroes. An exception to this is when the file has been opened in ‘direct_io’ mode, in which case the return value of the read system call will reflect the return value of this operation. fh will contain the value set by the open method, or will be undefined if the open method didn’t set any value.

flags: these are the file flags, such as O_SYNC. Only supported with ABI >= 7.9 lock_owner: only supported with ABI >= 7.9

fn write(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_offset: i64,
_data: &[u8],
_write_flags: u32,
_flags: i32,
_lock_owner: Option<u64>,
reply: ReplyWrite
)

Write data. Write should return exactly the number of bytes requested except on error. An exception to this is when the file has been opened in ‘direct_io’ mode, in which case the return value of the write system call will reflect the return value of this operation. fh will contain the value set by the open method, or will be undefined if the open method didn’t set any value.

write_flags: will contain FUSE_WRITE_CACHE, if this write is from the page cache. If set, the pid, uid, gid, and fh may not match the value that would have been sent if write cachin is disabled flags: these are the file flags, such as O_SYNC. Only supported with ABI >= 7.9 lock_owner: only supported with ABI >= 7.9

fn flush(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_lock_owner: u64,
reply: ReplyEmpty
)

Flush method. This is called on each close() of the opened file. Since file descriptors can be duplicated (dup, dup2, fork), for one open call there may be many flush calls. Filesystems shouldn’t assume that flush will always be called after some writes, or that if will be called at all. fh will contain the value set by the open method, or will be undefined if the open method didn’t set any value. NOTE: the name of the method is misleading, since (unlike fsync) the filesystem is not forced to flush pending writes. One reason to flush data, is if the filesystem wants to return write errors. If the filesystem supports file locking operations (setlk, getlk) it should remove all locks belonging to ‘lock_owner’.

fn release(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_flags: i32,
_lock_owner: Option<u64>,
_flush: bool,
reply: ReplyEmpty
)

Release an open file. Release is called when there are no more references to an open file: all file descriptors are closed and all memory mappings are unmapped. For every open call there will be exactly one release call. The filesystem may reply with an error, but error values are not returned to close() or munmap() which triggered the release. fh will contain the value set by the open method, or will be undefined if the open method didn’t set any value. flags will contain the same flags as for open.

fn fsync(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_datasync: bool,
reply: ReplyEmpty
)

Synchronize file contents. If the datasync parameter is non-zero, then only the user data should be flushed, not the meta data.

fn opendir(
&mut self,
_req: &Request<'_>,
_ino: u64,
_flags: i32,
reply: ReplyOpen
)

Open a directory. Filesystem may store an arbitrary file handle (pointer, index, etc) in fh, and use this in other all other directory stream operations (readdir, releasedir, fsyncdir). Filesystem may also implement stateless directory I/O and not store anything in fh, though that makes it impossible to implement standard conforming directory stream operations in case the contents of the directory can change between opendir and releasedir.

fn readdir(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_offset: i64,
reply: ReplyDirectory
)

Read directory. Send a buffer filled using buffer.fill(), with size not exceeding the requested size. Send an empty buffer on end of stream. fh will contain the value set by the opendir method, or will be undefined if the opendir method didn’t set any value.

fn readdirplus(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_offset: i64,
reply: ReplyDirectoryPlus
)

Read directory. Send a buffer filled using buffer.fill(), with size not exceeding the requested size. Send an empty buffer on end of stream. fh will contain the value set by the opendir method, or will be undefined if the opendir method didn’t set any value.

fn releasedir(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_flags: i32,
reply: ReplyEmpty
)

Release an open directory. For every opendir call there will be exactly one releasedir call. fh will contain the value set by the opendir method, or will be undefined if the opendir method didn’t set any value.

fn fsyncdir(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_datasync: bool,
reply: ReplyEmpty
)

Synchronize directory contents. If the datasync parameter is set, then only the directory contents should be flushed, not the meta data. fh will contain the value set by the opendir method, or will be undefined if the opendir method didn’t set any value.

fn statfs(&mut self, _req: &Request<'_>, _ino: u64, reply: ReplyStatfs)

Get file system statistics.

fn setxattr(
&mut self,
_req: &Request<'_>,
_ino: u64,
_name: &OsStr,
_value: &[u8],
_flags: i32,
_position: u32,
reply: ReplyEmpty
)

Set an extended attribute.

fn getxattr(
&mut self,
_req: &Request<'_>,
_ino: u64,
_name: &OsStr,
_size: u32,
reply: ReplyXattr
)

Get an extended attribute. If size is 0, the size of the value should be sent with reply.size(). If size is not 0, and the value fits, send it with reply.data(), or reply.error(ERANGE) if it doesn’t.

fn listxattr(
&mut self,
_req: &Request<'_>,
_ino: u64,
_size: u32,
reply: ReplyXattr
)

List extended attribute names. If size is 0, the size of the value should be sent with reply.size(). If size is not 0, and the value fits, send it with reply.data(), or reply.error(ERANGE) if it doesn’t.

fn removexattr(
&mut self,
_req: &Request<'_>,
_ino: u64,
_name: &OsStr,
reply: ReplyEmpty
)

Remove an extended attribute.

fn access(
&mut self,
_req: &Request<'_>,
_ino: u64,
_mask: i32,
reply: ReplyEmpty
)

Check file access permissions. This will be called for the access() system call. If the ‘default_permissions’ mount option is given, this method is not called. This method is not called under Linux kernel versions 2.4.x

fn create(
&mut self,
_req: &Request<'_>,
_parent: u64,
_name: &OsStr,
_mode: u32,
_umask: u32,
_flags: i32,
reply: ReplyCreate
)

Create and open a file. If the file does not exist, first create it with the specified mode, and then open it. Open flags (with the exception of O_NOCTTY) are available in flags. Filesystem may store an arbitrary file handle (pointer, index, etc) in fh, and use this in other all other file operations (read, write, flush, release, fsync). There are also some flags (direct_io, keep_cache) which the filesystem may set, to change the way the file is opened. See fuse_file_info structure in <fuse_common.h> for more details. If this method is not implemented or under Linux kernel versions earlier than 2.6.15, the mknod() and open() methods will be called instead.

fn getlk(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_lock_owner: u64,
_start: u64,
_end: u64,
_typ: i32,
_pid: u32,
reply: ReplyLock
)

Test for a POSIX file lock.

fn setlk(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_lock_owner: u64,
_start: u64,
_end: u64,
_typ: i32,
_pid: u32,
_sleep: bool,
reply: ReplyEmpty
)

Acquire, modify or release a POSIX file lock. For POSIX threads (NPTL) there’s a 1-1 relation between pid and owner, but otherwise this is not always the case. For checking lock ownership, ‘fi->owner’ must be used. The l_pid field in ‘struct flock’ should only be used to fill in this field in getlk(). Note: if the locking methods are not implemented, the kernel will still allow file locking to work locally. Hence these are only interesting for network filesystems and similar.

fn bmap(
&mut self,
_req: &Request<'_>,
_ino: u64,
_blocksize: u32,
_idx: u64,
reply: ReplyBmap
)

Map block index within file to block index within device. Note: This makes sense only for block device backed filesystems mounted with the ‘blkdev’ option

fn ioctl(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_flags: u32,
_cmd: u32,
_in_data: &[u8],
_out_size: u32,
reply: ReplyIoctl
)

control device

fn fallocate(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_offset: i64,
_length: i64,
_mode: i32,
reply: ReplyEmpty
)

Preallocate or deallocate space to a file

fn lseek(
&mut self,
_req: &Request<'_>,
_ino: u64,
_fh: u64,
_offset: i64,
_whence: i32,
reply: ReplyLseek
)

Reposition read/write file offset

fn copy_file_range(
&mut self,
_req: &Request<'_>,
_ino_in: u64,
_fh_in: u64,
_offset_in: i64,
_ino_out: u64,
_fh_out: u64,
_offset_out: i64,
_len: u64,
_flags: u32,
reply: ReplyWrite
)

Copy the specified range from the source inode to the destination inode